Move some components from common to libs/shared
