+++ es/es-discovery-ec2/build.gradle
@@ -1,6 +1,7 @@
 apply from: "$rootDir/gradle/javaModule.gradle"
 
 dependencies {
+    implementation project(':libs:shared')
     implementation project(':es:es-server')
     implementation "com.amazonaws:aws-java-sdk-ec2:${versions.aws}"
 }
