+++ es/es-server/src/main/java/org/elasticsearch/action/ActionListener.java
@@ -19,10 +19,10 @@
 
 package org.elasticsearch.action;
 
+import io.crate.common.CheckedSupplier;
 import org.elasticsearch.ExceptionsHelper;
 import org.elasticsearch.common.CheckedConsumer;
 import org.elasticsearch.common.CheckedFunction;
-import org.elasticsearch.common.CheckedSupplier;
 
 import java.util.ArrayList;
 import java.util.List;
