+++ libs/shared/src/main/java/io/crate/common/StringUtils.java
