+++ libs/shared/src/main/java/io/crate/common/TriConsumer.java
