+++ libs/shared/src/test/java/io/crate/common/StringUtilsTest.java
